rename revision folder to forms
